refactor(file-preview): organize page and tab hook
