+++ src/renderer/components/FilePreview/hooks/__tests__/useOpenFilePreviewTab.test.tsx
