+++ src/renderer/components/FilePreview/hooks/useOpenFilePreviewTab.ts
